+++ viteclientts/src/utils/store.tsx
@@ -1,7 +1,16 @@
 import { create } from 'zustand';
 import { UserState } from '../types/types';
 
-export const useUserStore = create<UserState>(set => ({
-    role: null,
-    setRole: (role) => set({ role })
-}))
+const useUserStore = create<UserState>((set) => ({
+    role: localStorage.getItem('userRole'), // Can be string or null
+    setRole: (role: string | null) => {
+        if (role === null) {
+            localStorage.removeItem('userRole'); // Remove the item if role is null
+        } else {
+            localStorage.setItem('userRole', role); // Persist the role
+        }
+        set({ role });
+    }
+}));
+
+export { useUserStore };